occupancy_component/requirements.test.txt
pytest==7.1.2
pytest-cov==3.0.0
pytest-homeassistant-custom-component==0.11.20
